Makefile: make check target depends on all-local
Without this patch "make check" won't build all the binaries prior to test
run.
This is, most probably, because TESTS are defined only in subdir.
And most probably there is a better way to fix this deps (for example, move
TESTS from tests/Makefile.am to the topmost Makefile.am)

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>