generator: abort patch creation if ".rodata" differs
Regular binaries are not "pic", and accesses to ".rodata" specified by
absolute address.
It mean, that _any_ change in ".text" size will shift ".rodata" start address
and, in turn, all the references in to it in ".text" will be different like
this:

424c7f:       b9 60 9d 49 00          mov $0x499d60,%ecx
424c7f:       b9 c0 9d 49 00          mov $0x499dc0,%ecx

There is not much we can do with it. But if ".rodata" sections content are
equal in two binaries, then we can assume, that these lines are the same:
1) value we put on register is within ".rodata" segment range and
2) Offset within ".rodata" segment is equal in old and new binary.

and if new function contains only such lines, then it can be considered as
unchanged.
This approach will help in 99,9% cases. There is a very low probability, that
this number will be not the address on ".rodata" but some number to calculate.
But it's very unlikely, that other code lines won't change in this case.
And it they change, then this function is a modified one anyway.

This is what we start with. More issues and code pieces because of ".rodata"
will follow.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>