tests: call old caller in loop while patching
It allows to test stack unwind.
Unfortunatelly, this testing is not garanteed to be correct.
Current statistics shows, that there is less that one catch of a process in a
wrong place per one tests run.
And there can be a situation, that a test can fail, because will ca caught in
a wrong palce for 5 times in a row.
So, it's not a best way to test stack rollback, but I don't have any better
idea so far.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>