tests: remotve printf from test_combined
For some reason test wrapper can hang after patching to this binary on write
of a huge string full of "PASS" words, received from patched process (note,
that "caller" helper is now called in loop).
So, let's drop this print. It's not really required and was just another PLT
entry to update on pathing. Let's use getpid() instead to keep the complexity.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>