patcher: rework extern symbol search
Looks like it doesn't matter, which type the symbol is: global or weak.
First occurence is used anyway.
There is a special rule for patched symbol: if defined in patch, it has to be
searched _only_ until old library. If not found, patched version has to be
used.
It mimics the logic of runnig the process with the new library (where the
patch is applied to the library binary already). In this case symbol will be
found not later than in the old library.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>