generator: check function for minimal size
Define this size as 8 bytes. This is a paranoya check.
This size is small enought (look like functoin size on x86_64 can never be
less than 16 bytes), but in the same time this check garantees, that function
has enough space for jump instrustion (which is 5 bytes on x86_64).
Also, 8 bytes is the size of ptrace operation, which will allow to simplify
patcher logic while reading and writing jumps and code.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>