patcher: compare real function bytes to correponding patch jump code on revert
Do this instead of "applied" flag in function jump.
This will allow to revert patch, which is partially covered by another patch.
IOW, if patch "A" jump instruction was overwritten by patch "B", it has to be
skipped on patch "A" revert. And this can be done only, if we compare code.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>