patcher: collect process needed list by using service link_map addresses
This is the only correct way to do so.
Because "NEEDED" tag is path, but not soname.
And this path may be a link (it this case mapping won't have "NEEDED" tag
value in it's name) or even deleted.
IOW, there is no reliable way to match "NEEDED" tag value with process
mapping.
NOTE: process "NEEDED" tag it not required to match library "SONAME" tag.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>