patcher: add strict argument to bt_check_range
In case of backtrace_check_func() we don't really case about first command in
function on stack, because it will be replaced by jump (if this function is
going to be patched), so this case can be optimized a bit by checking address
space in between borders.
But in case of backtrace_check_range() the check has to be strict, because
this function is intended to be used on patch revert, and thus all the region
will be unmapped.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>