tests: "const_var" test added
This test is aimed to test, that constant variale is patched properly.
First of all, this varibale will be optimed away (or at least it's expected).
IOW, there won't be any varibale - only value, returned from function.
Second, it means, that patched version has to return new value, but not the
old one.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>