tests: convert "global_var_{addr}" test into "manual" and "auto" versions
This is tempotary patch. Current solution with "vzpatch_" prefix in global
variable tests is wrong by design.
Prefix "vzpatch_" should be used only for static symbols.
While global symbols have to be resolved by generator as is and stored in
patch info separately.
Once it's done, "auto" tests has to be become the one for both modes.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>