service: send fd in mmap request instead of path
This is more appropriate approach.
Service mmap request is aimed to reflect mmap (2), which relies on descriptor.
But what is much more improtant, is that target process can have not right to
lookup or open patch file.
Currently, patch has to be placed somewhere in the process file system area
with proper (read) permissions to be able to patch it.
That's not good, especially when process is executed by some low-priveleged
user in it's own home.
Next patch will add patch fd sending to process, which will be used instead of
directly opened one.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>