patcher: remove process_release_at() helper
This helper is not very useful. To use ptrace to access process address space,
it has to be stopped.
Thus, releasin process and stopping it several times during patching is not
very usefull and not used right now.
Let's drop this feature and call plugin command recevier manually.
This is, BTW, a less dangerous approach.
Also, this will allow to use upstream libcompel instead of our own homebrew
version.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>