Makefile: convert generic LDFLAGS definition into test library LDFLAGS
Generic definition affects all the the libraries (including nsb plugin, which
is bad).
So, let's convert it into the test library LDFLAGS (this is needed for testing
to generate shared library upon build, and not upon installation).

Note: without this patch rpm builder fails to find versioned version of the
plugin library.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>