testrunner: wait to process to be served by dynamic linker
There is a race window between process start and check for mapping by Build ID
in it: testrunner can be faster than dymanic linker (i.e. mapping is not yet
ready).
Looks like it takes about millisecond for dynamic linker to mmap all the
binaries, but let's wait up to 50 millisecond with 1 millisecond step.
It should be enough for parallel test harness.

Signed-off-by: Stanislav Kinsburskiy <skinsbursky@virtuozzo.com>