generator: fix code identity checking
Previously we fed offset into relocation map only if byte values at this
offset were different. However, first bytes of the values may be the
same. If 'P' is offset, first difference may be found at 'P+1', and
this offset is not present in the relocation map. To fix this, use
relocation map regardless of data values.

Signed-off-by: Eugene Kolomeetz <klm@virtuozzo.com>