generator: handle multiple debug info keys for the same variable
It may so happen that the same variable can be described by multiple
debuginfo keys. This happens when corresponding symbol is placed in
common block. Consider such example:
-- a.c --
int var;

int *f(void)
{
        return &var;
}
-- b.c --
int var;

int *g(void)
{
        return &var;
}
-- c.c --
static int var;

int *h(void)
{
        return &var;
}
--
'var' in a.c & b.c is placed to common block:
	15: 0000000000000004      4 OBJECT  GLOBAL DEFAULT   COMMON var
and so, DIE for it is present in both files. Due to yet another 'var' in
c.c, generator is forced to use debuginfo, and it fails on reversing
debuginfo_key -> address map. To handle this, let's find all debuginfo
keys which map to given address in patch ELF, and use them to resolve
addresses in target ELF. Addresses should be the same for all keys.

Signed-off-by: Eugene Kolomeetz <klm@virtuozzo.com>